Memoize the get_contained_object_spec(), validation_spec(), and allowed_params() methods.  This can give a big speed boost when methods are called repeatedly, for example when using factory methods. All memoization caches are cleared when valid_params() or contained_objects() is called.  The only remaining known pitfall in the caching is that a class that dynamically changes its @ISA will probably mess things up.
Use the qr// format for regexes in the all_specs() method.

Improve the docs for allowed_params() - it's a class method (not an
instance method), and it accepts a list of arguments that can affect
the return value.

Don't copy as many hashes internally.  Pass by reference.

Version bump to 0.06.